Update deploy.yml
This commit updates the deploy.yml file to include the correct AWS S3 bucket name for deployment. The previous configuration used a placeholder ("your-bucket-name"), which needed to be replaced with the actual name of the S3 bucket where the project files will be deployed.
By updating the bucket name in the deployment workflow, It ensures that the project is deployed to the correct S3 bucket, allowing it to be accessible to users. This change is necessary for maintaining the integrity of the deployment process and ensuring that the latest version of the project is available to users.

No further changes were made in this commit. The deployment workflow remains unchanged, with only the bucket name updated.